refactor: split out request-to-input.go from checkhttp.go
